:wrench: Restructure project (#9)
* :wrench: Factor out jwt & errors code into separate modules

* :wrench: 'config/main.json' -> 'config.main.js'

* :wrench: Refactor 'routes/api.js' into seperate controllers

* :bug: Fix import bug in 'test/tests.js'

* :arrow_right: Move 'controllers' to 'routes/controllers'
