+++ src/validation/errors.js
@@ -1,0 +1,5 @@
+export default {
+  invalidToken: new Error('Invalid token'),
+  userNotFound: new Error('User not found'),
+  notAdmin: new Error('Not admin'),
+};
